// frontend/src/lib/api/types/ConnectionJump.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type ConnectionJump = { id: number, map_id: number, connection_id: number | null, character_id: number | null, character_name: string | null, ship_type_id: number | null, ship_type_name: string | null, 
/**
 * Hull mass in kg.
 */
mass: number, is_manual: boolean, from_solar_system_id: number, to_solar_system_id: number, created_at: string, };